Changes names of files subdirectories to be more durable.
The files are typically seen by the students, and upon specific instruction
by the teacher, so the context is usually there and does not need to be
reinforced or clarified by lesson number. Lesson numbers may still change,
and I'd hate these to get out of sync. Also, I'm shortening the names to
minimize typing for the students (who may not all know the command line
perks of auto-completion).

This also brings the README files in the directories up to date.